Use integer type for system plugins

+++ src/effects.c
@@ -716,10 +716,10 @@ static pthread_mutex_t g_hmi_mutex;
 
 /* system plugins */
 static int g_compressor_mode = 0;
-static float g_compressor_release = 100.0f;
+static int g_compressor_release = 100;
 static int g_noisegate_channel = 0;
 static int g_noisegate_decay = 10;
-static float g_noisegate_threshold = -60.0f;
+static int g_noisegate_threshold = -60;
 static gate_t g_noisegate;
 
 static const char* const g_bypass_port_symbol = BYPASS_PORT_SYMBOL;
@@ -1468,7 +1468,7 @@ static void* HMIClientThread(void* arg)
                 monitor_client_setup_compressor(g_compressor_mode, g_compressor_release);
                 break;
             case sys_serial_event_type_compressor_release:
-                g_compressor_release = atof(msg);
+                g_compressor_release = atoi(msg);
                 monitor_client_setup_compressor(g_compressor_mode, g_compressor_release);
                 break;
             case sys_serial_event_type_noisegate_channel:
@@ -1478,16 +1478,16 @@ static void* HMIClientThread(void* arg)
                 g_noisegate_decay = atoi(msg);
                 gate_update(&g_noisegate, g_sample_rate, 10, 1,
                             g_noisegate_decay, 1,
-                            g_noisegate_threshold, g_noisegate_threshold - 20.0f);
+                            g_noisegate_threshold, g_noisegate_threshold - 20);
                 break;
             case sys_serial_event_type_noisegate_threshold:
-                g_noisegate_threshold = atof(msg);
+                g_noisegate_threshold = atoi(msg);
                 gate_update(&g_noisegate, g_sample_rate, 10, 1,
                             g_noisegate_decay, 1,
-                            g_noisegate_threshold, g_noisegate_threshold - 20.0f);
+                            g_noisegate_threshold, g_noisegate_threshold - 20);
                 break;
             case sys_serial_event_type_pedalboard_gain:
-                monitor_client_setup_volume(atof(msg));
+                monitor_client_setup_volume(atoi(msg));
                 break;
             default:
                 break;
